requirements.txt
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
SpeechRecognition
fastapi
uvicorn
google-genai
glcoud
python-dotenv
PyAudio
dateparser
pydantic
openai-whisper
sounddevice
scipy
soundfile